excourt_re/excourt_re-master/system/requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-socketio==5.11.1
python-engineio==4.9.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
flask-swagger-ui==5.21.0
black
isort
requests
